refactor(dashboard): move resource version artifacts to handler
Why this change was needed:
Resource version artifact orchestration should not rely on local imports to avoid dependency cycles.

What changed:
Move artifact creation into a dedicated ResourceVersionArtifactHandler and keep ResourceVersionHandler free of OpenAPI exporter dependencies.

Problem solved:
Preserves the biz-layer orchestration while keeping imports explicit and compliant with dashboard architecture rules.

--story=1069995498134630065
