+++ src/dashboard/apigateway/apigateway/tests/biz/resource_version/test_resource_version.py
@@ -26,6 +26,7 @@ from apigateway.apps.openapi.models import OpenAPIFileResourceSchemaVersion, Ope
 from apigateway.apps.support.models import GatewaySDK, ReleasedResourceDoc, ResourceDoc, ResourceDocVersion
 from apigateway.biz.resource import ResourceHandler
 from apigateway.biz.resource_version import ResourceVersionHandler
+from apigateway.biz.resource_version.artifacts import ResourceVersionArtifactHandler
 from apigateway.core.constants import StageStatusEnum
 from apigateway.core.models import Gateway, Release, ReleasedResource, ResourceVersion, Stage
 from apigateway.utils.time import now_datetime
@@ -73,7 +74,7 @@ class TestResourceVersionHandler:
         )
         G(ResourceDoc, gateway=fake_gateway, resource_id=fake_resource.id)
 
-        result = ResourceVersionHandler.create_resource_version_with_artifacts(
+        result = ResourceVersionArtifactHandler.create_resource_version_with_artifacts(
             gateway=fake_gateway,
             data={"version": "20260526120000", "comment": "release comment"},
             username="admin",
